PoC-demo/requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
Authlib @ file:///Users/alpha/Desktop/Coding%20Projects/Active%20Personal%20Projects/zauth/authlib
billiard==3.6.3.0
celery==4.3.0
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==3.0.8
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
pycparser==2.20
python-bitcoinrpc==1.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
Werkzeug==1.0.1
zipp==3.1.0
